new: support more icalls on linux.

+++ LeanCLR~/runtime/icalls/system_runtime_runtimeimports.cpp
@@ -74,7 +74,7 @@ RtResultVoid SystemRuntimeRuntimeImports::ecvt_s(uint8_t* buffer, int32_t size, 
     {
         RET_ERR(RtErr::Argument);
     }
-    std::strncpy(reinterpret_cast<char*>(buffer), str, size - 1);
+    std::strncpy(reinterpret_cast<char*>(buffer), str, static_cast<size_t>(size - 1));
     buffer[size - 1] = 0;
 #else
     int written = snprintf(reinterpret_cast<char*>(buffer), static_cast<size_t>(size), "%.*e", digits, value);
